requirements.txt
fastapi[all]
llama-index
llama-index-agent-openai
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file 
llama-index-storage-docstore-mongodb
llama-index-vector-stores-mongodb
pymongo
motor
python-dotenv
llama-index-retrievers-bm25
python-jose 
PyJWT
passlib
bcrypt